moving otel to an internal package
